Access Control in Laravel

+++ app/Http/Controllers/PostsContoller.php
@@ -8,6 +8,17 @@ use DB;
 use App\User;
 class PostsContoller extends Controller
 {
+    
+    
+        /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -79,6 +90,11 @@ class PostsContoller extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorize page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -113,6 +129,10 @@ class PostsContoller extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorize page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
